Выделены модули event-service и request-service. Несколько классов дописаны, несколько удалено как не используемые.

+++ common/src/main/java/ru/yandex/practicum/dto/events/UpdateEventUserRequest.java
@@ -21,7 +21,7 @@ public class UpdateEventUserRequest {
     private Long id;
     @Size(min = 20, max = 2000)
     private String annotation;
-    private int category;
+    private Long category;
     @Size(min = 20, max = 7000)
     private String description;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = DateConfig.FORMAT)
